人物关系挖掘-规则提取V2-全角括号过滤

+++ work/src/include/relation.cpp
@@ -127,7 +127,7 @@ string Relation::added_id_by_kg(const string target, const string name) {
 
 string filterSpecial(const string text, const string startMark, const string endMark)
 {
-    int start, end, length;
+    size_t start, end, length;
     string ret;
     if ((start = text.find(startMark)) != string::npos) {
         if ((end = text.find(endMark)) != string::npos) {
